[prelude] move extension stream.unwrap to prelude
