+++ kyo-prelude/shared/src/test/scala/kyo/StreamTest.scala
@@ -1165,6 +1165,15 @@ class StreamTest extends Test:
 
             assert(someChoices == Chunk(1, 3))
         }
+
+        "extension" in {
+            val tick = Var.updateDiscard[Int](_ + 1)
+
+            val stream = tick.andThen(Stream.range(-1, 2)).unwrap
+
+            val n: Int = stream.foreach(_ => tick).handle(_.andThen(Var.get[Int]), Var.run(0)).eval
+            assert(n == 4)
+        }
     }
 
 end StreamTest
